Update the "concurrent requests" block
This has been left undone since 36e542d4.

The paragraph on what can already be expected of proxies is now grouped
with the rest of the "As it has always been" statements.

The actual concurrent operations section is now much shorter because
there is not really anything new introduced here.