Avoid actual request tag recycling when a security association spans network endpoints
The endpoint definition of 7252 was written with DTLS in mind where a
security association is only ever alive with one IP/port. (Session
resumption AFAICT counts as new association).

Closing 19, because while we can't change RFC7959, we now give it enough
consideration that this specification does not stand in the way of
anyone interpreting RFC7959 in a more flexible way and switches
endpoints during a transfer (as long as the security context stands).

Closes: https://github.com/core-wg/echo-request-tag/issues/19